ecommerce_order_service/requirements.txt
asgiref==3.11.1
attrs==26.1.0
awsebcli==3.27.1
bcrypt==5.0.0
blessed==1.27.0
botocore==1.42.88
cement==2.10.14
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
colorama==0.4.6
cryptography==46.0.7
decorator==5.2.1
Deprecated==1.3.1
Django==6.0.4
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
fabric==3.2.2
idna==3.11
inflection==0.5.1
invoke==3.0.3
jmespath==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
mysqlclient==2.2.8
packaging==26.0
paramiko==4.0.0
pathspec==0.12.1
pycparser==3.0
PyJWT==2.12.1
PyNaCl==1.6.2
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.3
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
semantic-version==2.10.0
setuptools==82.0.1
six==1.17.0
sqlparse==0.5.5
termcolor==2.5.0
uritemplate==4.2.0
urllib3==2.6.3
wcwidth==0.2.14
wrapt==2.1.2
